+++ packages/common/src/index.ts
@@ -2,3 +2,4 @@ export * from './js.js';
 export * from './time.js';
 export * from './utils/index.js';
 export * from './types/index.js';
+export * from './ethereum.js';