+++ packages/client/CHANGELOG.md
@@ -1,5 +1,16 @@
 # fuzd-client
 
+## 0.9.1
+
+### Patch Changes
+
+- fixes + bestTime
+- Updated dependencies
+  - fuzd-chain-protocol@0.9.1
+  - fuzd-scheduler@0.9.1
+  - fuzd-executor@0.9.1
+  - fuzd-common@0.9.1
+
 ## 0.9.0
 
 ### Minor Changes