+++ packages/common/CHANGELOG.md
@@ -1,5 +1,11 @@
 # fuzd-executor
 
+## 0.9.1
+
+### Patch Changes
+
+- fixes + bestTime
+
 ## 0.9.0
 
 ### Minor Changes